refactor(view) : 자판기 출력에 맞게 함수들 위치 이동
UserView -> VendingMachineView
