+++ src/main/java/vendingmachine/view/VendingMachineView.java
@@ -14,4 +14,21 @@ public class VendingMachineView {
 
 	}
 
+	public void printRemainOfInsertedMoney(int remainMoney) {
+		System.out.println("\n" + OutputMessage.PRINT_USER_MONEY + remainMoney + "원");
+	}
+
+	public void printChangesComment(int remainMoney) {
+		printRemainOfInsertedMoney(remainMoney);
+		System.out.println("\n" + OutputMessage.PRINT_CHANGES);
+	}
+
+	public void printChanges(int countCoin, Coin coin) {
+
+		if (countCoin != 0) {
+			System.out.println(coin.getAmount() + "원 - " + countCoin + "개");
+		}
+
+	}
+
 }
